% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gate_seurat_obj}
\alias{gate_seurat_obj}
\title{gate_seurat_obj}
\format{
A list containing x,y coordinates for one gate
}
\usage{
data(gate_seurat_obj)
}
\description{
Coordinates for a gate interactively drawn using tidygate
}
\keyword{datasets}
