% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sce_obj_UMAP3}
\alias{sce_obj_UMAP3}
\title{sce_obj_UMAP3}
\format{
A sce object.
}
\usage{
data(sce_obj_UMAP3)
}
\description{
A sce dataset of single cell RNA sequencing data with 3 UMAP dimesions
}
\keyword{datasets}
